PCSM-322: Simplify unsuccessful index finalization
