+++ pcsm/catalog/catalog.go
@@ -3,12 +3,9 @@
 package catalog
 
 import (
-	"bytes"
 	"context"
 	"encoding/hex"
 	"math"
-	"reflect"
-	"slices"
 	"sort"
 	"strings"
 	"sync"
@@ -177,56 +174,6 @@ type indexCatalogEntry struct {
 func (i indexCatalogEntry) Unsuccessful() bool {
 	return i.Failed || i.Incomplete || i.Inconsistent
 }
-
-// IndexUnsuccessfulType describes why an index ended up unsuccessful at finalize time.
-type IndexUnsuccessfulType string
-
-const (
-	// IndexFailed means the index failed to create on the target cluster.
-	IndexFailed IndexUnsuccessfulType = "failed"
-	// IndexIncomplete means the index was being built on the source when replication observed it.
-	IndexIncomplete IndexUnsuccessfulType = "incomplete"
-	// IndexInconsistent means the index was inconsistent across shards on the source cluster.
-	IndexInconsistent IndexUnsuccessfulType = "inconsistent"
-)
-
-// UnsuccessfulIndex describes an index that did not complete cleanly during replication
-// and was not recovered during finalize.
-type UnsuccessfulIndex struct {
-	Namespace string
-	Name      string
-	Keys      bson.Raw
-	Type      IndexUnsuccessfulType
-	Reason    string
-}
-
-type finalizeIndexDecisionInput struct {
-	storedSpec     *mdb.IndexSpecification
-	sourceSpec     *mdb.IndexSpecification
-	sourceCheckErr error
-	originalType   IndexUnsuccessfulType
-	inProgress     bool
-	inconsistent   bool
-}
-
-type finalizeIndexDecision struct {
-	reportType IndexUnsuccessfulType
-	reason     string
-	recreate   bool
-}
-
-type finalizeSourceIndexInput struct {
-	namespace    Namespace
-	index        indexCatalogEntry
-	originalType IndexUnsuccessfulType
-}
-
-const (
-	finalizeReasonNoLongerPresent         = "no longer present on source"
-	finalizeReasonSourceIndexBuilding     = "index is still building on one or more source shards"
-	finalizeReasonSourceIndexInconsistent = "index is missing on one or more source shards"
-	finalizeReasonSourceSpecChanged       = "source spec changed"
-)
 
 // NewCatalog creates a new Catalog.
 func NewCatalog(source, target *mongo.Client, sourceVer mdb.ServerVersion) *Catalog {
@@ -995,9 +942,9 @@ func (c *Catalog) UUIDMap() UUIDMap {
 //     after target index-creation failures are recreated unconditionally.
 //     - Incomplete and Inconsistent entries are rechecked against the source.
 //     They are recreated only when the source index is neither building nor
-//     inconsistent and still exists with the specification recorded during
-//     cloning. Source-check failures, unresolved, missing, or changed indexes,
-//     and target recreation failures are reported with the corresponding reason.
+//     inconsistent and still exists, using its current source specification.
+//     Source-check failures, unresolved or missing indexes, and target
+//     recreation failures are reported with the corresponding reason.
 func (c *Catalog) Finalize(ctx context.Context) []UnsuccessfulIndex {
 	lg := log.Ctx(ctx)
 
@@ -1131,336 +1078,6 @@ func (c *Catalog) Finalize(ctx context.Context) []UnsuccessfulIndex {
 	})
 
 	return report
-}
-
-// finalizeUnsuccessfulIndexes finalizes indexes that were unsuccessful during
-// replication. Failed entries keep the historical always-recreate behavior.
-// Incomplete/Inconsistent entries are rechecked on source before recreate.
-// The recheck and recreate are not atomic: replication is paused, but source
-// DDL is not fenced, so the source-still-valid decision is best-effort.
-func (c *Catalog) finalizeUnsuccessfulIndexes(ctx context.Context) []UnsuccessfulIndex {
-	lg := log.Ctx(ctx)
-	lg.Info("Finalizing unsuccessful indexes")
-
-	var report []UnsuccessfulIndex
-
-	for db, colls := range c.Databases {
-		for coll, collEntry := range colls.Collections {
-			for _, index := range collEntry.Indexes {
-				if !index.Unsuccessful() {
-					continue // skip successful indexes
-				}
-
-				originalType := index.unsuccessfulType()
-				if originalType != IndexFailed {
-					decision := c.decideFinalizeSourceIndex(ctx, finalizeSourceIndexInput{
-						namespace:    Namespace{Database: db, Collection: coll},
-						index:        index,
-						originalType: originalType,
-					})
-					if !decision.recreate {
-						lg.Warnf("Index %s on %s.%s was not recreated during finalize: %s",
-							index.Name, db, coll, decision.reason)
-
-						report = append(report, UnsuccessfulIndex{
-							Namespace: db + "." + coll,
-							Name:      index.Name,
-							Keys:      index.KeysDocument,
-							Type:      decision.reportType,
-							Reason:    decision.reason,
-						})
-
-						continue
-					}
-
-					lg.Infof("Index %s on %s.%s is valid on source, trying to recreate it",
-						index.Name, db, coll)
-				} else {
-					lg.Infof("Index %s on %s.%s failed to create during replication, trying to recreate it",
-						index.Name, db, coll)
-				}
-
-				err := runWithRetry(ctx, func(ctx context.Context) error {
-					err := c.target.Database(db).RunCommand(ctx, bson.D{
-						{"createIndexes", coll},
-						{"indexes", bson.A{index.IndexSpecification}},
-					}).Err()
-
-					return errors.Wrapf(err, "recreate index %s.%s.%s", db, coll, index.Name)
-				})
-				if err != nil {
-					lg.Warnf("Failed to recreate unsuccessful index %s on %s.%s: %v",
-						index.Name, db, coll, err)
-
-					report = append(report, UnsuccessfulIndex{
-						Namespace: db + "." + coll,
-						Name:      index.Name,
-						Keys:      index.KeysDocument,
-						Type:      originalType,
-						Reason:    err.Error(),
-					})
-
-					continue
-				}
-
-				lg.Infof("Recreated index %s on %s.%s", index.Name, db, coll)
-
-				c.lock.Lock()
-				c.addIndexesToCatalog(ctx, db, coll, []indexCatalogEntry{{IndexSpecification: index.IndexSpecification}})
-				c.lock.Unlock()
-			}
-		}
-	}
-
-	return report
-}
-
-func (i indexCatalogEntry) unsuccessfulType() IndexUnsuccessfulType {
-	switch {
-	case i.Failed:
-		return IndexFailed
-	case i.Incomplete:
-		return IndexIncomplete
-	case i.Inconsistent:
-		return IndexInconsistent
-	default:
-		return ""
-	}
-}
-
-func (c *Catalog) decideFinalizeSourceIndex(
-	ctx context.Context,
-	params finalizeSourceIndexInput,
-) finalizeIndexDecision {
-	index := params.index
-	originalType := params.originalType
-
-	input := finalizeIndexDecisionInput{
-		storedSpec:   index.IndexSpecification,
-		originalType: originalType,
-	}
-
-	switch originalType {
-	case IndexFailed:
-		return decideFinalizeUnsuccessfulIndex(input)
-
-	case IndexIncomplete:
-		inProgress, err := c.sourceIndexInProgress(ctx, params.namespace, index.Name)
-		if err != nil {
-			input.sourceCheckErr = err
-
-			return decideFinalizeUnsuccessfulIndex(input)
-		}
-
-		input.inProgress = inProgress
-		if inProgress {
-			return decideFinalizeUnsuccessfulIndex(input)
-		}
-
-		inconsistent, err := c.sourceIndexInconsistent(ctx, params.namespace, index.Name)
-		if err != nil {
-			input.sourceCheckErr = err
-
-			return decideFinalizeUnsuccessfulIndex(input)
-		}
-
-		input.inconsistent = inconsistent
-		if inconsistent {
-			return decideFinalizeUnsuccessfulIndex(input)
-		}
-
-	case IndexInconsistent:
-		inconsistent, err := c.sourceIndexInconsistent(ctx, params.namespace, index.Name)
-		if err != nil {
-			input.sourceCheckErr = err
-
-			return decideFinalizeUnsuccessfulIndex(input)
-		}
-
-		input.inconsistent = inconsistent
-		if inconsistent {
-			return decideFinalizeUnsuccessfulIndex(input)
-		}
-
-		inProgress, err := c.sourceIndexInProgress(ctx, params.namespace, index.Name)
-		if err != nil {
-			input.sourceCheckErr = err
-
-			return decideFinalizeUnsuccessfulIndex(input)
-		}
-
-		input.inProgress = inProgress
-		if inProgress {
-			return decideFinalizeUnsuccessfulIndex(input)
-		}
-	}
-
-	indexes, err := mdb.ListIndexes(
-		ctx,
-		c.source,
-		params.namespace.Database,
-		params.namespace.Collection,
-	)
-	if err != nil {
-		input.sourceCheckErr = errors.Wrap(err, "source")
-
-		return decideFinalizeUnsuccessfulIndex(input)
-	}
-
-	input.sourceSpec = findIndexByName(indexes, index.Name)
-
-	return decideFinalizeUnsuccessfulIndex(input)
-}
-
-func (c *Catalog) sourceIndexInProgress(ctx context.Context, namespace Namespace, name string) (bool, error) {
-	inProgress, err := mdb.ListInProgressIndexBuilds(ctx, c.source, namespace.Database, namespace.Collection)
-	if err != nil {
-		return false, errors.Wrap(err, "list source in-progress index builds")
-	}
-
-	return slices.Contains(inProgress, name), nil
-}
-
-func (c *Catalog) sourceIndexInconsistent(ctx context.Context, namespace Namespace, name string) (bool, error) {
-	inconsistent, err := mdb.ListInconsistentIndexes(ctx, c.source, namespace.Database, namespace.Collection)
-	if err != nil {
-		return false, errors.Wrap(err, "list source inconsistent indexes")
-	}
-
-	return findIndexByName(inconsistent, name) != nil, nil
-}
-
-func findIndexByName(indexes []*mdb.IndexSpecification, name string) *mdb.IndexSpecification {
-	idx := slices.IndexFunc(indexes, func(spec *mdb.IndexSpecification) bool {
-		return spec != nil && spec.Name == name
-	})
-	if idx == -1 {
-		return nil
-	}
-
-	return indexes[idx]
-}
-
-func decideFinalizeUnsuccessfulIndex(input finalizeIndexDecisionInput) finalizeIndexDecision {
-	if input.originalType == IndexFailed {
-		return finalizeIndexDecision{recreate: true}
-	}
-
-	if input.sourceCheckErr != nil {
-		return finalizeIndexDecision{
-			reportType: input.originalType,
-			reason:     input.sourceCheckErr.Error(),
-		}
-	}
-
-	switch input.originalType { //nolint:exhaustive // IndexFailed handled by the guard above
-	case IndexIncomplete:
-		if input.inProgress {
-			return finalizeIndexDecision{
-				reportType: IndexIncomplete,
-				reason:     finalizeReasonSourceIndexBuilding,
-			}
-		}
-
-		if input.inconsistent {
-			return finalizeIndexDecision{
-				reportType: IndexInconsistent,
-				reason:     finalizeReasonSourceIndexInconsistent,
-			}
-		}
-
-	case IndexInconsistent:
-		if input.inconsistent {
-			return finalizeIndexDecision{
-				reportType: IndexInconsistent,
-				reason:     finalizeReasonSourceIndexInconsistent,
-			}
-		}
-
-		if input.inProgress {
-			return finalizeIndexDecision{
-				reportType: IndexIncomplete,
-				reason:     finalizeReasonSourceIndexBuilding,
-			}
-		}
-	}
-
-	if input.sourceSpec == nil {
-		return finalizeIndexDecision{
-			reportType: input.originalType,
-			reason:     finalizeReasonNoLongerPresent,
-		}
-	}
-
-	if !indexCreateSpecsEqual(input.storedSpec, input.sourceSpec) {
-		return finalizeIndexDecision{
-			reportType: input.originalType,
-			reason:     finalizeReasonSourceSpecChanged,
-		}
-	}
-
-	return finalizeIndexDecision{recreate: true}
-}
-
-func indexCreateSpecsEqual(stored, source *mdb.IndexSpecification) bool {
-	if stored == nil || source == nil {
-		return stored == nil && source == nil
-	}
-
-	// Compare only createIndexes-relevant fields. Ignore server-managed or
-	// non-create metadata: v, ns, and clustered. Clustered is a collection
-	// property surfaced on IndexSpecification, not an index recreated through
-	// createIndexes during finalize.
-	return bytes.Equal(stored.KeysDocument, source.KeysDocument) &&
-		ptrEqual(stored.Unique, source.Unique) &&
-		ptrEqual(stored.PrepareUnique, source.PrepareUnique) &&
-		ptrEqual(stored.Sparse, source.Sparse) &&
-		ptrEqual(stored.Hidden, source.Hidden) &&
-		ptrEqual(stored.ExpireAfterSeconds, source.ExpireAfterSeconds) &&
-		bytes.Equal(stored.Collation, source.Collation) &&
-		bytes.Equal(stored.StorageEngine, source.StorageEngine) &&
-		indexOptionValueEqual(stored.Weights, source.Weights) &&
-		ptrEqual(stored.DefaultLanguage, source.DefaultLanguage) &&
-		ptrEqual(stored.LanguageOverride, source.LanguageOverride) &&
-		ptrEqual(stored.TextVersion, source.TextVersion) &&
-		indexOptionValueEqual(stored.WildcardProjection, source.WildcardProjection) &&
-		indexOptionValueEqual(stored.PartialFilterExpression, source.PartialFilterExpression) &&
-		ptrEqual(stored.Bits, source.Bits) &&
-		ptrEqual(stored.Min, source.Min) &&
-		ptrEqual(stored.Max, source.Max) &&
-		ptrEqual(stored.GeoIdxVer, source.GeoIdxVer)
-}
-
-// indexOptionValueEqual compares option sub-documents by marshaled BSON
-// (field-order sensitive; relies on stable same-source key order).
-func indexOptionValueEqual(left, right any) bool {
-	if left == nil || right == nil {
-		return left == nil && right == nil
-	}
-
-	leftRaw, leftOK := marshalIndexOptionValue(left)
-	rightRaw, rightOK := marshalIndexOptionValue(right)
-	if leftOK && rightOK && bytes.Equal(leftRaw, rightRaw) {
-		return true
-	}
-
-	return reflect.DeepEqual(left, right)
-}
-
-func marshalIndexOptionValue(value any) (bson.Raw, bool) {
-	switch v := value.(type) {
-	case bson.Raw:
-		return v, true
-	case []byte:
-		return bson.Raw(v), true
-	}
-
-	raw, err := bson.Marshal(value)
-	if err != nil {
-		return nil, false
-	}
-
-	return raw, true
 }
 
 // doModifyIndexOption modifies an index property in the target MongoDB.
@@ -1809,12 +1426,4 @@ func runWithRetry(
 	fn func(context.Context) error,
 ) error {
 	return mdb.RunWithRetry(ctx, fn, mdb.DefaultRetryInterval, mdb.DefaultMaxRetries) //nolint:wrapcheck
-}
-
-func ptrEqual[T comparable](left, right *T) bool {
-	if left == nil || right == nil {
-		return left == right
-	}
-
-	return *left == *right
 }
